+++ terraform_gcp_infra/modules/project-base/main.tf
@@ -14,7 +14,10 @@ resource "google_project" "this" {
   billing_account     = var.billing_account
   labels              = var.labels
   auto_create_network = false
-  deletion_policy     = var.deletion_policy
+
+  lifecycle {
+    prevent_destroy = var.prevent_destroy
+  }
 }
 
 # 1) 필수 API